+++ src/core/debug.go
@@ -5,7 +5,6 @@ package core
 
 import (
 	"fmt"
-
 	"net/http"
 	_ "net/http/pprof"
 	"os"
