+++ src/core/nodeinfo.go
@@ -11,9 +11,6 @@ import (
 
 	iwt "github.com/Arceliar/ironwood/types"
 	"github.com/Arceliar/phony"
-
-	//"github.com/yggdrasil-network/yggdrasil-go/src/crypto"
-
 	"github.com/yggdrasil-network/yggdrasil-go/src/version"
 )
